% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFun.R
\name{BasicHeatmap}
\alias{BasicHeatmap}
\title{Draw Basic heatmaps with not annotation}
\usage{
BasicHeatmap(
  x,
  normlize = "log2",
  scale = "row",
  color_min = "cyan",
  color_mid = "black",
  color_max = "gold",
  break_min,
  break_mid,
  break_max,
  cluster_type = "hclust",
  cluster_row = FALSE,
  cluster_col = FALSE,
  km_col,
  km_row,
  km_repeat_time = 100,
  cluster_row_method,
  cluster_row_distance,
  cluster_col_method,
  cluster_col_distance,
  show_rowname = T,
  show_colname = T
)
}
\arguments{
\item{x}{Data for heatmap after cleaning.}

\item{normlize}{Normlization method for log2,log10,log2 x+1 ,log10 x+1}

\item{scale}{Z-score scaled data input, rowscale, colscale and none, Default = "row".}

\item{color_min}{Color for minmal values.}

\item{color_mid}{Color for middle values.}

\item{color_max}{Color for max values.}

\item{break_min}{min values.}

\item{break_mid}{mid values.}

\item{break_max}{max values.}

\item{cluster_type}{Hierarchical clustering or k-means.}

\item{cluster_row}{For hclust, cluster row or not, default = FALSE.}

\item{cluster_col}{For hclust, cluster col or not, default = FALSE.}

\item{km_col}{For k-means, number of expected group of column.}

\item{km_row}{For k-means, number of expected group of column.}

\item{km_repeat_time}{Avoid different results of k-means each time, Default = 100.}

\item{cluster_row_method}{see hclust.}

\item{cluster_row_distance}{see hclust.}

\item{cluster_col_method}{see hclust.}

\item{cluster_col_distance}{see hclust.}

\item{show_rowname}{Show rownames or not.}

\item{show_colname}{Show colnames or not.}
}
\value{
a CompexHeatmap result
}
\description{
Draw Basic heatmaps with not annotation
}
\references{
https://jokergoo.github.io/ComplexHeatmap-reference/book/
}
\author{
Shawn Wang <http://www.shawnlearnbioinfo.top>
}
